dynamic_cli

Module prelude

Prelude module for quickly importing essential types

This module re-exports the most commonly used types and traits, allowing you to import everything with a single use statement.

Example

use dynamic_cli::prelude::*;

// Now you have access to:
// - ExecutionContext, downcast_ref, downcast_mut
// - CommandHandler
// - DynamicCliError, Result
// - CommandRegistry
// - ParsedCommand, CliParser, ReplParser
// - validate_file_exists, validate_file_extension, validate_range
// - Common config types (ArgumentType, CommandsConfig)
// - CliBuilder, CliApp
// - Utility functions (parse_int, parse_bool, is_blank, etc.)

Re-exports

pub use crate::context::downcast_mut;

pub use crate::context::downcast_ref;

pub use crate::context::ExecutionContext;

pub use crate::executor::CommandHandler;

pub use crate::error::DynamicCliError;

pub use crate::error::Result;

pub use crate::config::schema::ArgumentType;

pub use crate::config::schema::CommandsConfig;

pub use crate::registry::CommandRegistry;

pub use crate::parser::CliParser;

pub use crate::parser::ParsedCommand;

pub use crate::parser::ReplParser;

pub use crate::validator::validate_file_exists;

pub use crate::validator::validate_file_extension;

pub use crate::validator::validate_range;

pub use crate::interface::CliInterface;

pub use crate::interface::ReplInterface;

pub use crate::builder::CliApp;

pub use crate::builder::CliBuilder;

pub use crate::help::DefaultHelpFormatter;

pub use crate::help::HelpFormatter;

pub use crate::utils::detect_type;

pub use crate::utils::is_blank;

pub use crate::utils::normalize;

pub use crate::utils::parse_bool;

pub use crate::utils::parse_float;

pub use crate::utils::parse_int;